MAINT: Place confirm close hook in ui utils
