+++ frontend/src/utils/form.ts
@@ -3,7 +3,6 @@ import {
   type AnyFormApi,
   createFormHookContexts,
 } from "@tanstack/react-form";
-import { useEffect, useState } from "react";
 
 import type { Smda } from "#client";
 import type { OptionProps } from "#components/form/field";
@@ -39,51 +38,6 @@ export function findOptionValueInNameUuidArray<T extends NameUuidType>(
   const result = array.filter((element) => String(element.uuid) === value);
 
   return result.length === 1 ? result[0] : undefined;
-}
-
-export function useConfirmClose({
-  enable,
-  determineRequiresConfirmation,
-  onCloseConfirmed,
-}: {
-  enable: boolean;
-  determineRequiresConfirmation: () => boolean;
-  onCloseConfirmed: () => void;
-}) {
-  const [confirmCloseDialogOpen, setConfirmCloseDialogOpen] = useState(false);
-
-  useEffect(() => {
-    if (!enable) {
-      setConfirmCloseDialogOpen(false);
-    }
-  }, [enable]);
-
-  const closeDialogs = () => {
-    setConfirmCloseDialogOpen(false);
-    onCloseConfirmed();
-  };
-
-  const handleCloseRequest = () => {
-    if (determineRequiresConfirmation()) {
-      setConfirmCloseDialogOpen(true);
-    } else {
-      closeDialogs();
-    }
-  };
-
-  const handleDecision = (confirm: boolean) => {
-    if (confirm) {
-      closeDialogs();
-    } else {
-      setConfirmCloseDialogOpen(false);
-    }
-  };
-
-  return {
-    confirmCloseDialogOpen,
-    handleCloseRequest,
-    handleDecision,
-  };
 }
 
 /**
